06_FitStep/requirements.txt
rich==13.7.1
openai>=2.0.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
chromadb>=0.4.0
